Check if board is solvable before solving it
This is needed because the algorithm does not check if the board is
valid, only checking if a value can be inserted on each cell. This was
an issue on boards with the same value on the same row/column/square.

Also display a feedback to the user on the status